+++ Block/Zone/Request.php
@@ -5,9 +5,12 @@
  * Do not edit or add to this file if you wish to upgrade this module
  * to newer versions in the future.
  */
+declare(strict_types=1);
+
 namespace Smile\DebugToolbar\Block\Zone;
 
 use Magento\Framework\App\Request\Http as MagentoRequest;
+use Magento\Framework\App\RequestInterface;
 
 /**
  * Zone for Debug Toolbar Block
@@ -26,7 +29,7 @@ class Request extends AbstractZone
     /**
      * @inheritdoc
      */
-    public function getCode()
+    public function getCode(): string
     {
         return 'request';
     }
@@ -34,7 +37,7 @@ class Request extends AbstractZone
     /**
      * @inheritdoc
      */
-    public function getTitle()
+    public function getTitle(): string
     {
         return 'Request';
     }
@@ -45,7 +48,7 @@ class Request extends AbstractZone
      * @param MagentoRequest $request
      * @return $this
      */
-    public function setRequest(MagentoRequest $request)
+    public function setRequest(MagentoRequest $request): Request
     {
         $this->request = $request;
 
@@ -55,7 +58,7 @@ class Request extends AbstractZone
     /**
      * @inheritdoc
      */
-    public function getRequest()
+    public function getRequest(): RequestInterface
     {
         return $this->request;
     }
@@ -65,8 +68,8 @@ class Request extends AbstractZone
      *
      * @return string
      */
-    public function getControllerClassName()
+    public function getControllerClassName(): string
     {
-        return $this->helperData->getValue('controller_classname');
+        return (string) $this->helperData->getValue('controller_classname');
     }
 }