+++ Block/Zone/Response.php
@@ -5,6 +5,8 @@
  * Do not edit or add to this file if you wish to upgrade this module
  * to newer versions in the future.
  */
+declare(strict_types=1);
+
 namespace Smile\DebugToolbar\Block\Zone;
 
 use Magento\Framework\HTTP\PhpEnvironment\Response as MagentoResponse;
@@ -26,7 +28,7 @@ class Response extends AbstractZone
     /**
      * @inheritdoc
      */
-    public function getCode()
+    public function getCode(): string
     {
         return 'response';
     }
@@ -34,7 +36,7 @@ class Response extends AbstractZone
     /**
      * @inheritdoc
      */
-    public function getTitle()
+    public function getTitle(): string
     {
         return 'Response';
     }
@@ -45,7 +47,7 @@ class Response extends AbstractZone
      * @param MagentoResponse $response
      * @return $this
      */
-    public function setResponse(MagentoResponse $response)
+    public function setResponse(MagentoResponse $response): Response
     {
         $this->response = $response;
 
@@ -57,7 +59,7 @@ class Response extends AbstractZone
      *
      * @return MagentoResponse
      */
-    public function getResponse()
+    public function getResponse(): MagentoResponse
     {
         return $this->response;
     }
@@ -67,7 +69,7 @@ class Response extends AbstractZone
      *
      * @return string
      */
-    public function getFullPageCacheMode()
+    public function getFullPageCacheMode(): string
     {
         return $this->helperData->getFullPageCacheMode();
     }
@@ -77,7 +79,7 @@ class Response extends AbstractZone
      *
      * @return string[]
      */
-    public function getEsiUrlList()
+    public function getEsiUrlList(): array
     {
         $list = [];