+++ Helper/Observer.php
@@ -5,6 +5,8 @@
  * Do not edit or add to this file if you wish to upgrade this module
  * to newer versions in the future.
  */
+declare(strict_types=1);
+
 namespace Smile\DebugToolbar\Helper;
 
 use Magento\Framework\App\Helper\AbstractHelper;
@@ -28,7 +30,7 @@ class Observer extends AbstractHelper
      *
      * @param string $eventName
      */
-    public function initEventStat($eventName)
+    public function initEventStat(string $eventName): void
     {
         $eventName = strtolower($eventName);
         if (!array_key_exists($eventName, $this->eventStats)) {
@@ -53,8 +55,12 @@ class Observer extends AbstractHelper
      * @param string $observerInstance
      * @param bool $observerDisabled
      */
-    public function initObserverStat($eventName, $observerName, $observerInstance, $observerDisabled)
-    {
+    public function initObserverStat(
+        string $eventName,
+        string $observerName,
+        string $observerInstance,
+        bool $observerDisabled
+    ): void {
         $eventName = strtolower($eventName);
         if (!array_key_exists($observerName, $this->eventStats[$eventName]['observers'])) {
             $this->eventStats[$eventName]['observers'][$observerName] = [
@@ -78,7 +84,7 @@ class Observer extends AbstractHelper
      * @param string $eventName
      * @param float $deltaTime
      */
-    public function addEventStat($eventName, $deltaTime)
+    public function addEventStat(string $eventName, float $deltaTime): void
     {
         $eventName = strtolower($eventName);
         $usage = $this->eventStats[$eventName];
@@ -96,7 +102,7 @@ class Observer extends AbstractHelper
      * @param string $observerName
      * @param float $deltaTime
      */
-    public function addObserverStat($eventName, $observerName, $deltaTime)
+    public function addObserverStat(string $eventName, string $observerName, float $deltaTime): void
     {
         $eventName = strtolower($eventName);
         $usage = $this->eventStats[$eventName]['observers'][$observerName];
@@ -112,7 +118,7 @@ class Observer extends AbstractHelper
      *
      * @return array
      */
-    public function getEventStats()
+    public function getEventStats(): array
     {
         return $this->eventStats;
     }