+++ Plugin/App/Http.php
@@ -5,6 +5,8 @@
  * Do not edit or add to this file if you wish to upgrade this module
  * to newer versions in the future.
  */
+declare(strict_types=1);
+
 namespace Smile\DebugToolbar\Plugin\App;
 
 use Magento\Framework\App\Http as MagentoHttp;
@@ -47,7 +49,7 @@ class Http
      * @return array
      * @SuppressWarnings(PHPMD.UnusedFormalParameter)
      */
-    public function beforeLaunch(MagentoHttp $subject)
+    public function beforeLaunch(MagentoHttp $subject): array
     {
         if ($this->helperConfig->isEnabled()) {
             $this->helperData->startTimer('app_http');