+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## WIP
+
+- Add type hinting and strict types
+
 ## [4.0.3] - 2019-11-26
 [4.0.3]: https://github.com/Smile-SA/magento2-module-debug-toolbar/compare/4.0.2...4.0.3